+++ packages/react/src/hooks/web3inboxClient.ts
@@ -63,37 +63,34 @@ export const useWeb3InboxClient = () => {
 
 export const useW3iAccount = () => {
   const client = useWeb3InboxClient();
-  const [identityKey, setIdentityKey] = useState<string | null>(null);
+  const [isRegistered, setIsRegistered] = useState<boolean>(false);
+  const [isRegistering, setIsRegistering] = useState<boolean>(false);
 
-  const { account } = useClientState();
+  const { account, registration } = useClientState();
 
   const setAccount = useCallback(
-    (account: string) => {
-      console.log("Attempting to Setting the account to the client");
+    async (account: string) => {
       if (client) {
-        console.log("Setting the account to the client");
-        client.setAccount(account);
+        return client.setAccount(account);
       }
     },
-    [client]
+    [client, account]
   );
 
-  // Account for the changing of the account
   useEffect(() => {
-    if (!account) {
-      setIdentityKey(null);
-    }
-  }, [account]);
+    const registrationStatus = registration? registration.account === account : false;
+    setIsRegistered(registrationStatus)
+  }, [account, registration])
 
   const register = useCallback(
     async (onSign: (m: string) => Promise<string>) => {
       if (client && account) {
+	setIsRegistering(true);
         const identity = await client.register({
           account,
           onSign,
         });
-
-        setIdentityKey(identity);
+	setIsRegistering(false)
         return identity;
       }
 
@@ -102,5 +99,12 @@ export const useW3iAccount = () => {
     [client, account]
   );
 
-  return { account, setAccount, register, identityKey };
+  return {
+    account,
+    setAccount,
+    register,
+    isRegistering,
+    isRegistered,
+    identityKey: isRegistered && registration ? registration.identity : undefined
+  };
 };