+++ packages/react/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @web3inbox/widget-react
 
+## 0.6.22
+
+### Patch Changes
+
+- Expose isRegistering
+- Updated dependencies
+  - @web3inbox/core@0.0.19
+
+## 0.6.21
+
+### Patch Changes
+
+- Update identity key on register
+- Updated dependencies
+  - @web3inbox/core@0.0.18
+
+## 0.6.20
+
+### Patch Changes
+
+- Make identity keys reactive state
+- Updated dependencies
+  - @web3inbox/core@0.0.17
+
+## 0.6.19
+
+### Patch Changes
+
+- Switch core to peer dep of widget
+- Updated dependencies
+  - @web3inbox/core@0.0.16
+
 ## 0.6.18
 
 ### Patch Changes